move movieEntity and watchlistMovieEntity to data layer
