+++ prog-2-teamAF/src/main/java/com/teamAF/app/Data/MovieEntity.java
@@ -1,7 +1,8 @@
-package com.teamAF.app.Model;
+package com.teamAF.app.Data;
 
 import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.table.DatabaseTable;
+import com.teamAF.app.Model.Movie;
 
 import java.util.ArrayList;
 import java.util.Arrays;
