+++ prog-2-teamAF/src/main/java/com/teamAF/app/Data/MovieRepository.java
@@ -3,8 +3,6 @@ package com.teamAF.app.Data;
 import com.j256.ormlite.dao.Dao;
 import com.j256.ormlite.stmt.QueryBuilder;
 import com.teamAF.app.Model.Movie;
-import com.teamAF.app.Model.MovieEntity;
-import com.teamAF.app.Model.WatchlistMovieEntity;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
