+++ prog-2-teamAF/src/test/java/com/teamAF/app/Data/MovieRepositoryTest.java
@@ -2,7 +2,6 @@ package com.teamAF.app.Data;
 
 import com.j256.ormlite.dao.Dao;
 import com.teamAF.app.Model.Movie;
-import com.teamAF.app.Model.MovieEntity;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
