+++ prog-2-teamAF/src/test/java/com/teamAF/app/Model/MovieEntityTest.java
@@ -1,5 +1,6 @@
 package com.teamAF.app.Model;
 
+import com.teamAF.app.Data.MovieEntity;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
